granted auth in all routes

+++ server/src/routes/memories.ts
@@ -3,8 +3,15 @@ import { z } from 'zod'
 import { prisma } from '../lib/prisma'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request: FastifyRequest) => {
+    await request.jwtVerify
+  })
+
+  app.get('/memories', async (request: FastifyRequest) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -17,21 +24,27 @@ export async function memoriesRoutes(app: FastifyInstance) {
     }))
   })
 
-  app.get('/memories/:id', async (request: FastifyRequest) => {
-    const paramsSchema = z.object({
-      id: z.string().uuid(),
-    })
+  app.get(
+    '/memories/:id',
+    async (request: FastifyRequest, reply: FastifyReply) => {
+      const paramsSchema = z.object({
+        id: z.string().uuid(),
+      })
 
-    const { id } = paramsSchema.parse(request.params)
+      const { id } = paramsSchema.parse(request.params)
 
-    const memory = await prisma.memory.findUniqueOrThrow({
-      where: {
-        id,
-      },
-    })
+      const memory = await prisma.memory.findUniqueOrThrow({
+        where: {
+          id,
+        },
+      })
 
-    return memory
-  })
+      if (!memory.isPublic && memory.userId !== request.user.sub)
+        return reply.status(401).send()
+
+      return memory
+    },
+  )
 
   app.post(
     '/memories',
@@ -49,11 +62,11 @@ export async function memoriesRoutes(app: FastifyInstance) {
           content,
           coverUrl,
           isPublic,
-          userId: '06fed0ee-5064-4363-8a2b-c35c3d086f57',
+          userId: request.user.sub,
         },
       })
 
-      return reply.send(memory).status(201)
+      return reply.status(201).send(memory)
     },
   )
 
@@ -73,7 +86,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
       const { id } = paramsSchema.parse(request.params)
       const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-      const memory = await prisma.memory.update({
+      let memory = await prisma.memory.findUniqueOrThrow({
+        where: {
+          id,
+        },
+      })
+
+      if (memory.userId !== request.user.sub) return reply.status(401).send()
+
+      memory = await prisma.memory.update({
         where: { id },
         data: {
           content,
@@ -95,13 +116,21 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
       const { id } = paramsSchema.parse(request.params)
 
+      const memory = await prisma.memory.findUniqueOrThrow({
+        where: {
+          id,
+        },
+      })
+
+      if (memory.userId !== request.user.sub) return reply.status(401).send()
+
       await prisma.memory.delete({
         where: {
           id,
         },
       })
 
-      return reply.status(200)
+      return reply.status(200).send()
     },
   )
 }
